+++ index.php
@@ -14,16 +14,15 @@ require 'vendor/autoload.php';
 require_once 'apps/init.php';
 
 // use naming or namespace
-use MiniMvc\Core\App;
-use MiniMvc\Core\Route;
+use MiniMvc\Apps\Core\Bootstraping\App;
+use MiniMvc\Apps\Core\Bootstraping\Routes;
 
 /**
  *  Initialisasi Class App untuk menjalankan File App
  * 	App/core/App.php
  */
-
+$Route = new Routes;
 $App = new App;
-$Route = new Route;
 
 // var_dump($Route->info());
